refactor: fix data races and split api.go into logical files
